+++ packages/blackborne/src/components/Steps/index.ts
@@ -1,2 +1,7 @@
 export { Step, Steps } from './Steps';
-export type { StepProps, StepsProps, StepStatus } from './Steps';
+export type {
+  StepProps,
+  StepsProps,
+  StepStatus,
+  StepTitlePlacement
+} from './Steps';